+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## develop
 - update default values from v2 of API
+- fix bug for tiered, time-of-use demand rates
 
 
 ## v0.13.0